Add more validations. Throw ResponseStatusException accordingly to the type of the method that failed. add Utils -> for implementing common methods which everybody have to use.

+++ src/main/java/recipeapplication/database/UserFirebaseCrud.java
@@ -1,15 +1,17 @@
 package recipeapplication.database;
 
-import TransferObjects.UserEditTO;
+import Utils.FinalsStringsExceptions;
+import Utils.StringUtils;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.cloud.firestore.*;
-import net.bytebuddy.implementation.bytecode.Throw;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 
 import com.google.cloud.firestore.Firestore;
 
+import org.springframework.web.server.ResponseStatusException;
 import recipeapplication.components.User;
 
 import java.util.Map;
@@ -30,17 +32,24 @@ public class UserFirebaseCrud implements FirebaseCrud<User> {
     }
 
     @Override
-    public void create(User user) {
-
-        //Todo - more validations
-        this.firestore.collection(collection).document(user.getId()).set(user);
+    public User create(User user) throws Exception {
+        if (!isAllUserDetailsValid(user)) {
+            this.firestore.collection(collection).document(user.getId()).set(user);
+        } else {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not create the user, please verify that all the required parameters are filled correctly");
+        }
+        return user;
     }
 
     @Override
     public User update(User user) throws Exception {
+        if (user.getFavoriteRecipes() == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The edited user has not favoriteRecipes. Can't update his details ");
+        }
+
         User userFromDb = read(user.getId());
         if (userFromDb == null) {
-            throw new Exception("User does not exist");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, FinalsStringsExceptions.NO_USER_FOUND);
         } else {
             user.setRole(userFromDb.getRole());
             user.setName(userFromDb.getName());
@@ -50,34 +59,50 @@ public class UserFirebaseCrud implements FirebaseCrud<User> {
         return user;
     }
 
-
     @Override
-    public User delete(String document) {
-        //Todo - more validations (empty id ..)
+    public User delete(String document) throws Exception {
         /*Fb delete does not return an object, just details about the deletion time so we have to read the object in order to return it back*/
         User user = read(document);
-        /*delete does not fail if a document doesn't exist*/
+        if (user == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, FinalsStringsExceptions.NO_USER_FOUND);
+        }
+
         this.firestore.collection(collection).document(document).delete();
         return user;
     }
 
     @Override
-    public User read(String document) {
-        //Todo - more validations  (empty id ..)
-        DocumentReference userRef = this.firestore.collection(collection).document(document);
-        User userPojo = null;
-        try {
-            DocumentSnapshot userSnapshot = userRef.get().get();
-            Map<String, Object> userMap = userSnapshot.getData();
-            if (userMap != null) {
-                /*Use jackson in order to convert the map to an user instance*/
-                userPojo = mapper.convertValue(userMap, User.class);
+    public User read(String document) throws Exception {
+
+        if (!StringUtils.isNullOrEmptyTrimmed(document)) {
+            DocumentReference userRef = this.firestore.collection(collection).document(document);
+            User userPojo = null;
+            try {
+                DocumentSnapshot userSnapshot = userRef.get().get();
+                Map<String, Object> userMap = userSnapshot.getData();
+                if (userMap != null) {
+                    /*Use jackson in order to convert the map to an user instance*/
+                    userPojo = mapper.convertValue(userMap, User.class);
+                } else {
+                    throw new ResponseStatusException(HttpStatus.NOT_FOUND, FinalsStringsExceptions.NO_USER_FOUND);
+                }
+            } catch (Exception e) {
+                if (e instanceof ResponseStatusException) {
+                    throw e;
+                } else {
+                    System.err.println("Could not read the user");
+                }
             }
-        } catch (Exception e) {
-            System.err.println("Could not delete the user");
+            return userPojo;
+        } else {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not fetch the user, please verify that all the required parameters are filled correctly");
         }
-        return userPojo;
     }
 
+    private boolean isAllUserDetailsValid(User user) {
+        return StringUtils.isNullOrEmptyTrimmed(user.getId()) || user.getName() == null ||
+                StringUtils.isNullOrEmptyTrimmed(user.getName().getFirst()) || StringUtils.isNullOrEmptyTrimmed(user.getName().getLast())
+                || user.getRole() == null;
+    }
 
 }
